# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "这是软件杯B1赛题",
        "title": "龙芯服务 API",
        "termsOfService": "http://swagger.io/terms/ 使用条款",
        "contact": {
            "name": "LoveSong.",
            "url": "无",
            "email": "3130250166@qq.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "39.99.139.249:8080",
    "paths": {
...
        "/monitor/add": {
            "post": {
                "description": "添加监控系统信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "monitor"
                ],
                "summary": "添加监控系统信息",
                "parameters": [
                    {
                        "description": "需要用户 hardware是检测的硬件 detail是如果是io或net才需要传要不是all要不是单个名称 Threshold是百分比用户设定的 up是io或net的上行或上传 down反过来 notifyEmail是要通知邮箱可以多个传过来要用,分割",
                        "name": "req",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/model.Monitor"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/monitor/del": {
            "get": {
                "description": "按照id删除监控",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "monitor"
                ],
                "summary": "按照id删除监控",
                "parameters": [
                    {
                        "type": "string",
                        "description": "后端传的id",
                        "name": "id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/monitor/sel": {
            "post": {
                "description": "查询监控系统信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "monitor"
                ],
                "summary": "查询监控系统信息",
                "parameters": [
                    {
                        "description": "分页",
                        "name": "page",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/model.PageInfo"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
...
        "/sys/info": {
            "get": {
                "description": "服务器硬件信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "system"
                ],
                "summary": "服务器硬件信息",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/sysMes.LocalInfo"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/sys/io": {
            "get": {
                "description": "获取io监控信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "system"
                ],
                "summary": "获取io监控信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "该值表示要监控的磁盘名称All表示全部",
                        "name": "name",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/condition.DiskIo"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/sys/name": {
            "get": {
                "description": "获取io或net名称",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "system"
                ],
                "summary": "获取io或net名称",
                "parameters": [
                    {
                        "type": "string",
                        "description": "io 或者是 net",
                        "name": "name",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/condition.NetOrDiskName"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/sys/net": {
            "get": {
                "description": "获取网络监控信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "system"
                ],
                "summary": "获取网络监控信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "该值表示要监控的网络名称All表示全部",
                        "name": "name",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/condition.Flow"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/sys/static": {
            "get": {
                "description": "获取服务器状态",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "system"
                ],
                "summary": "获取服务器状态",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/sysStatic.SysStatus"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        }
    },
    "definitions": {
        "condition.DiskIo": {
            "type": "object",
            "properties": {
                "ioCount": {
                    "type": "integer"
                },
                "ioReadBytes": {
                    "type": "integer"
                },
                "ioReadTime": {
                    "type": "integer"
                },
                "ioWriteBytes": {
                    "type": "integer"
                },
                "ioWriterTime": {
                    "type": "integer"
                }
            }
        },
        "condition.Flow": {
            "type": "object",
            "properties": {
                "bytesRecv": {
                    "type": "integer"
                },
                "bytesSent": {
                    "type": "integer"
                },
                "recv": {
                    "type": "number"
                },
                "recvUnit": {
                    "type": "string"
                },
                "sent": {
                    "type": "number"
                },
                "sentUnit": {
                    "type": "string"
                }
            }
        },
        "condition.NetOrDiskName": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "model.Monitor": {
            "type": "object",
            "properties": {
                "createTime": {
                    "type": "string"
                },
                "createUser": {
                    "type": "string"
                },
                "detail": {
                    "type": "string"
                },
                "down": {
                    "type": "number"
                },
                "hardWare": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "notifyEmail": {
                    "type": "string"
                },
                "threshold": {
                    "type": "number"
                },
                "up": {
                    "type": "number"
                }
            }
        },
        "model.PageInfo": {
            "type": "object",
            "required": [
                "page",
                "pageSize"
            ],
            "properties": {
                "page": {
                    "type": "integer"
                },
                "pageSize": {
                    "type": "integer"
                }
            }
        },
...
        "sysMes.LocalInfo": {
            "type": "object",
            "properties": {
                "hostName": {
                    "type": "string"
                },
                "kernelV": {
                    "type": "string"
                },
                "platform": {
                    "type": "string"
                },
                "runTime": {
                    "type": "string"
                },
                "startTime": {
                    "type": "string"
                },
                "sysType": {
                    "type": "string"
                }
            }
        },
        "sysStatic.SysStatus": {
            "type": "object",
            "properties": {
                "available": {
                    "type": "number"
                },
                "cpuUsed": {
                    "type": "number"
                },
                "cpuUsedPercent": {
                    "type": "number"
                },
                "device": {
                    "type": "string"
                },
                "free": {
                    "type": "string"
                },
                "fstype": {
                    "type": "string"
                },
                "inodesFree": {
                    "type": "integer"
                },
                "inodesTotal": {
                    "type": "integer"
                },
                "inodesUsed": {
                    "type": "integer"
                },
                "inodesUsedPercent": {
                    "type": "number"
                },
                "load1": {
                    "type": "number"
                },
                "load15": {
                    "type": "number"
                },
                "load5": {
                    "type": "number"
                },
                "loadUsagePercent": {
                    "type": "number"
                },
                "mFree": {
                    "type": "number"
                },
                "mTotal": {
                    "type": "number"
                },
                "mUsed": {
                    "type": "number"
                },
                "mUsedPercent": {
                    "type": "number"
                },
                "model": {
                    "type": "string"
                },
                "numCPU": {
                    "type": "integer"
                },
                "numLogicCpu": {
                    "type": "integer"
                },
                "path": {
                    "type": "string"
                },
                "percentage": {
                    "type": "array",
                    "items": {
                        "type": "number"
                    }
                },
                "swapFree": {
                    "type": "number"
                },
                "swapPercent": {
                    "type": "number"
                },
                "swapTotal": {
                    "type": "number"
                },
                "swapUsed": {
                    "type": "number"
                },
                "total": {
                    "type": "string"
                },
                "used": {
                    "type": "string"
                },
                "usedPercent": {
                    "type": "number"
                }
            }
        }
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    },
...
}